+++ src/main/java/frc/robot/Robot.java
@@ -10,7 +10,7 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 
 import frc.robot.commands.LimelightCommand;
-// import frc.robot.subsystems.LimelightSubsystem;
+import frc.robot.subsystems.LimelightSubsystem;
 
 /**
  * The VM is configured to automatically run this class, and to call the functions corresponding to
@@ -97,8 +97,7 @@ public class Robot extends TimedRobot {
   @Override
   public void teleopPeriodic() {
     //runs limelight 
-    // LimelightCommand limelight = new LimelightCommand(limelightSubsystem);
-    LimelightCommand limelight = new LimelightCommand();
+    LimelightCommand limelight = new LimelightCommand(LimelightSubsystem);
     limelight.execute();
